requirements.txt
# HUBEX Backend Dependencies
alembic==1.17.2
asyncpg==0.31.0
httpx==0.28.1
bcrypt==3.2.2
email-validator==2.2.0
fastapi==0.125.0
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.11
pydantic[email]==2.12.5
pydantic-settings==2.12.0
python-dotenv==1.2.1
python-jose==3.5.0
python-json-logger==2.0.7
python-multipart==0.0.24
redis[asyncio]>=5.0,<6.0
SQLAlchemy==2.0.45
uvicorn[standard]==0.38.0
watchfiles==0.24.0
qrcode[svg]==8.0
requests>=2.31.0
